Move troi bits... to troi
