+++ troi/content_resolver/cli.py
@@ -11,13 +11,10 @@ from troi.content_resolver.database import Database
 from troi.content_resolver.model.recording import FileIdType
 from troi.content_resolver.subsonic import SubsonicDatabase
 from troi.content_resolver.metadata_lookup import MetadataLookup
-from troi.content_resolver.lb_radio import ListenBrainzRadioLocal
 from troi.content_resolver.utils import ask_yes_no_question
 from troi.content_resolver.top_tags import TopTags
 from troi.content_resolver.duplicates import FindDuplicates
-from troi.content_resolver.artist_search import LocalRecordingSearchByArtistService
-from troi.content_resolver.troi.periodic_jams import LocalPeriodicJams
-from troi.content_resolver.playlist import read_jspf_playlist, write_m3u_playlist, write_jspf_playlist
+from troi.content_resolver.playlist import write_m3u_playlist, write_jspf_playlist
 from troi.content_resolver.unresolved_recording import UnresolvedRecordingTracker
 from troi.playlist import PLAYLIST_TRACK_EXTENSION_URI
 
@@ -174,75 +171,6 @@ def subsonic(db_file):
 
 @click.command()
 @click.option("-d", "--db_file", help="Database file for the local collection", required=False, is_flag=False)
-@click.option('-t', '--threshold', default=.80)
-@click.option('-u', '--upload-to-subsonic', required=False, is_flag=True)
-@click.option('-m', '--save-to-m3u', required=False)
-@click.option('-j', '--save-to-jspf', required=False)
-@click.option('-y', '--dont-ask', required=False, is_flag=True, help="write playlist to m3u file")
-@click.argument('jspf_playlist')
-def playlist(db_file, threshold, upload_to_subsonic, save_to_m3u, save_to_jspf, dont_ask, jspf_playlist):
-    """ Resolve a JSPF file with MusicBrainz recording MBIDs to files in the local collection"""
-    db_file = db_file_check(db_file)
-    db = SubsonicDatabase(db_file, config)
-    db.open()
-    lbrl = ListenBrainzRadioLocal()
-    playlist = read_jspf_playlist(jspf_playlist)
-    lbrl.resolve_playlist(threshold, playlist)
-    output_playlist(db, playlist, upload_to_subsonic, save_to_m3u, save_to_jspf, dont_ask)
-
-
-@click.command()
-@click.option("-d", "--db_file", help="Database file for the local collection", required=False, is_flag=False)
-@click.option('-t', '--threshold', default=.80)
-@click.option('-u', '--upload-to-subsonic', required=False, is_flag=True)
-@click.option('-m', '--save-to-m3u', required=False)
-@click.option('-j', '--save-to-jspf', required=False)
-@click.option('-y', '--dont-ask', required=False, is_flag=True, help="write playlist to m3u file")
-@click.argument('mode')
-@click.argument('prompt')
-def lb_radio(db_file, threshold, upload_to_subsonic, save_to_m3u, save_to_jspf, dont_ask, mode, prompt):
-    """Use the ListenBrainz Radio engine to create a playlist from a prompt, using a local music collection"""
-    db_file = db_file_check(db_file)
-    db = SubsonicDatabase(db_file, config)
-    db.open()
-    r = ListenBrainzRadioLocal()
-    playlist = r.generate(mode, prompt, threshold)
-    try:
-        _ = playlist.playlists[0].recordings[0]
-    except (KeyError, IndexError, AttributeError):
-        db.metadata_sanity_check(include_subsonic=upload_to_subsonic)
-        return
-
-    output_playlist(db, playlist, upload_to_subsonic, save_to_m3u, save_to_jspf, dont_ask)
-
-
-@click.command()
-@click.option("-d", "--db_file", help="Database file for the local collection", required=False, is_flag=False)
-@click.option('-t', '--threshold', default=.80)
-@click.option('-u', '--upload-to-subsonic', required=False, is_flag=True, default=False)
-@click.option('-m', '--save-to-m3u', required=False)
-@click.option('-j', '--save-to-jspf', required=False)
-@click.option('-y', '--dont-ask', required=False, is_flag=True, help="write playlist to m3u file")
-@click.argument('user_name')
-def periodic_jams(db_file, threshold, upload_to_subsonic, save_to_m3u, save_to_jspf, dont_ask, user_name):
-    "Generate a periodic jams playlist"
-    db_file = db_file_check(db_file)
-    db = SubsonicDatabase(db_file, config)
-    db.open()
-
-    pj = LocalPeriodicJams(user_name, threshold)
-    playlist = pj.generate()
-    try:
-        _ = playlist.playlists[0].recordings[0]
-    except (KeyError, IndexError, AttributeError):
-        db.metadata_sanity_check(include_subsonic=upload_to_subsonic)
-        return
-
-    output_playlist(db, playlist, upload_to_subsonic, save_to_m3u, save_to_jspf, dont_ask)
-
-
-@click.command()
-@click.option("-d", "--db_file", help="Database file for the local collection", required=False, is_flag=False)
 @click.argument('count', required=False, default=250)
 def top_tags(db_file, count):
     "Display the top most used tags in the music collection. Useful for writing LB Radio tag prompts"
@@ -281,14 +209,11 @@ def unresolved(db_file):
 
 cli.add_command(create)
 cli.add_command(scan)
-cli.add_command(playlist)
 cli.add_command(cleanup)
 cli.add_command(metadata)
 cli.add_command(subsonic)
-cli.add_command(lb_radio)
 cli.add_command(top_tags)
 cli.add_command(duplicates)
-cli.add_command(periodic_jams)
 cli.add_command(unresolved)
